requirements.txt
fastapi
uvicorn
streamlit
openai
google-generativeai
faiss-cpu
pypdf
python-docx
python-dotenv
pydantic
requests
python-multipart
sentence-transformers
